fix (CreteConfigurationFileCommand): Checks if the configuration file already exists to prevent override

+++ app/Commands/CreateConfigurationFileCommand.php
@@ -2,8 +2,10 @@
 
 namespace Wtyd\GitHooks\App\Commands;
 
+// use Illuminate\Support\Facades\Storage;
 use LaravelZero\Framework\Commands\Command;
 use Wtyd\GitHooks\Utils\Printer;
+use Wtyd\GitHooks\Utils\Storage;
 
 class CreateConfigurationFileCommand extends Command
 {
@@ -23,20 +25,29 @@ class CreateConfigurationFileCommand extends Command
 
     public function handle()
     {
-        $root = getcwd();
-        $origin = "$root/vendor/wtyd/githooks/qa/githooks.dist.yml";
+        $origin = "vendor/wtyd/githooks/qa/githooks.dist.yml";
+        $destiny = "githooks.yml";
 
-        $destiny = "$root/githooks.yml";
-        // $this->printer->error("Failed to copy $origin to $destiny");
-        // $this->printer->success('Configuration file githooks.yml has been created in root path');
-        // exit();
-        return $this->copyFile($origin, $destiny);
+        if ($this->checkIfConfigurationFileExists()) {
+            return $this->copyFile($origin, $destiny);
+        } else {
+            return 1;
+        }
+    }
+
+    protected function checkIfConfigurationFileExists(): bool
+    {
+        if (Storage::exists('githooks.yml') || Storage::exists('qa/githooks.yml')) {
+            $this->printer->error('Configuration file githooks.yml already exists in root path');
+            return false;
+        }
+        return true;
     }
 
     protected function copyFile(string $origin, string $destiny): int
     {
         try {
-            if (copy($origin, $destiny) === false) {
+            if (!Storage::copy($origin, $destiny)) {
                 $this->printer->error("Failed to copy $origin to $destiny");
                 return 1;
             } else {
